Fixes a potential race condition from two callers
It was possible for two callers to hit an interesting race
condition where:

If two callers called into the runner at the same time,
both would enter the CoroutineRunner at the same point.

At this point one would claim the mutex while the other
waits. The caller A would check the active task, see that
there is none, and continue.

At this point caller A would create a new task. Caller B
would enter the mutex at the same point at which caller A
is creating a new task.

Caller A would then attempt to claim the lock again to
set itself as the active task. Caller B would be holding
the lock as it attempts to check for an active task, of
which there is none.  Once caller B leaves the lock, caller
A would claim it and set itself active, but caller B
would at this point be creating a new task and waiting
claim the lock to set itself active - thus creating a
situation where two calls to are running at once, both believing
that they are the active task.